[integration testing] switch to `integration_test` package as this is now the recommended way of performing integration_test. (test_driver is deprecated)
