+++ pubspec.yaml
@@ -73,7 +73,7 @@ dev_dependencies:
   dependency_validator: any
   flutter_test:
     sdk: flutter
-  flutter_driver:
+  integration_test:
     sdk: flutter
   test: any
