--- test_driver/app.dart
@@ -1,31 +1,0 @@
-import 'package:encointer_wallet/config.dart';
-import 'package:flutter/material.dart';
-
-import 'package:encointer_wallet/app.dart';
-import 'package:flutter_driver/driver_extension.dart';
-
-void main() {
-
-  // //
-  // Future<String> dataHandler(String msg) async {
-  //   switch (msg) {
-  //     case SETUP_STORE:
-  //       {
-  //         setupLocalStorage(globalAppStore);
-  //       }
-  //       break;
-  //     default:
-  //       break;
-  //   }
-  // }
-
-  enableFlutterDriverExtension();
-
-  // Call the `main()` function of the app, or call `runApp` with
-  // any widget you are interested in testing.
-  runApp(
-    WalletApp(Config(mockLocalStorage: true)),
-  );
-}
-
-const SETUP_STORE = 'setup_store';
